Sonoma Coast!
I've just finished up Sonoma Coast AVA, the line I've drawn for the
coast is in between the coastline and first contour line, so it doesn't
get every detail of the coastline - I figured that was okay. The
"approved maps" section has all the quads I used to replace the Sonoma
map.