Digitized Yamhill-Carlton
There is a strange part to this on the Fairdale Quadrangle map.  In (9) and (10) the west border crosses Panther Creek at the 123°17′30″ longitude line. This same line forms the west border of the entire region in (14), which splits the AVA into a north part and south part on each side of Panther Creek.  It is never shown this way by the Oregon Wine Board but is technically correct.