Madera History
Both revisions are minor edits, the first adding a county to the description and the second editing a typo (R 20E to R 21 E).  The first revision was made before the effective date of the AVA, so it has the same date appended to the ava_id, with an "_a" and a "_b" to separate between original and revision.