Grand Valley Touch-up
The border was slightly off in some sections and used an unnecessary amount of vertices in others so I edited these.  For the section that follows a wider river, I moved the border from the bank to the center of the river because a bank was not specified and the center requires a lower level of detail, less vertices.   Also fixed some typos in the attribute table.